requirements.txt
# ─── Backend ──────────────────────────────────────────────────
Django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
mysqlclient==2.2.4           # MySQL driver (or use PyMySQL below)
# PyMySQL==1.1.1             # Uncomment if mysqlclient fails to compile

# ─── AI / Vector DB ───────────────────────────────────────────
openai==1.30.1               # Works with LM Studio (local) and OpenAI Cloud
faiss-cpu==1.13.2
sentence-transformers==2.7.0
torch==2.3.0                 # Required by sentence-transformers (CPU only by default)

# ─── Scraper ──────────────────────────────────────────────────
selenium==4.20.0
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.3

# ─── Utilities ────────────────────────────────────────────────
Pillow==10.3.0
